Update README.md
Video Sharing Website

This project involves the design and development of a video-sharing website inspired by YouTube. It was built using HTML, CSS, and JavaScript.

Features  

Home Page: Includes a left sidebar, a banner, and a list of videos. Each video entry displays a thumbnail, title, and channel information.

Video Watch Page: Features a video player with detailed video and channel information, a list of comments, and a sidebar with suggested videos.

Responsive Design 

The website is designed to be responsive and functional on small-screen devices.

Technologies Used

HTML
CSS
JavaScript